moved failure callback classes to separate directory, added auth_tok to constructor for streamlining, added thoughts about modularizing notification modalities better
