+++ src/ingest-pipeline/airflow/dags/validate_upload.py
@@ -9,13 +9,13 @@ from airflow.configuration import conf as airflow_conf
 from airflow.operators.python import PythonOperator
 from airflow.exceptions import AirflowException
 from airflow.providers.http.hooks.http import HttpHook
-from airflow.utils.email import send_email
 
 from hubmap_operators.common_operators import (
     CreateTmpDirOperator,
     CleanupTmpDirOperator,
 )
 
+from error_catching.validate_upload_failure import ValidateUploadFailure
 from utils import (
     encrypt_tok,
     get_tmp_dir_path,
@@ -26,9 +26,9 @@ from utils import (
     get_preserve_scratch_resource,
     get_threads_resource,
 )
-from catch_pipeline_errors import FailureCallback, FailureCallbackException
 
 sys.path.append(airflow_conf.as_dict()["connections"]["SRC_PATH"].strip("'").strip('"'))
+
 from submodules import (
     ingest_validation_tools_upload,  # noqa E402
     ingest_validation_tools_error_report,
@@ -36,57 +36,6 @@ from submodules import (
 )
 
 sys.path.pop()
-
-
-class ValidateUploadFailure(FailureCallback):
-    # Should probably be importing custom exceptions rather than comparing strings
-    external_exceptions = [
-        "ValueError",
-        "PreflightError",
-        "ValidatorError",
-        "DirectoryValidationErrors",
-        "FileNotFoundError",
-    ]
-
-    def get_failure_email_template(
-        self,
-        formatted_exception=None,
-        external_template=False,
-        submission_data=None,
-        report_txt=False,
-    ):
-        if external_template:
-            if report_txt and submission_data:
-                subject = f"Your {submission_data.get('entity_type')} has failed!"
-                msg = f"""
-                    Error: {report_txt}
-                    """
-                return subject, msg
-        else:
-            if report_txt and submission_data:
-                subject = f"{submission_data.get('entity_type')} {self.uuid} has failed!"
-                msg = f"""
-                    Error: {report_txt}
-                    """
-                return subject, msg
-        return super().get_failure_email_template(formatted_exception)
-
-    def send_failure_email(self, **kwargs):
-        super().send_failure_email(**kwargs)
-        if "report_txt" in kwargs:
-            try:
-                created_by_user_email = self.submission_data.get("created_by_user_email")
-                subject, msg = self.get_failure_email_template(
-                    formatted_exception=None,
-                    external_template=True,
-                    submission_data=self.submission_data,
-                    **kwargs,
-                )
-                send_email(to=[created_by_user_email], subject=subject, html_content=msg)
-            except:
-                raise FailureCallbackException(
-                    "Failure retrieving created_by_user_email or sending email in ValidateUploadFailure."
-                )
 
 
 # Following are defaults which can be overridden later on
@@ -98,8 +47,8 @@ default_args = {
     "email_on_failure": False,
     "email_on_retry": False,
     "on_failure_callback": ValidateUploadFailure,
-    # "retries": 1,
-    "retries": 0,
+    "retries": 1,
+    # "retries": 0,
     "retry_delay": timedelta(minutes=1),
     "xcom_push": True,
     "queue": get_queue_resource("validate_upload"),
